requirements.txt
# ==============================================
# S-ACM - Smart Academic Content Management System
# Requirements File
# ==============================================

# Django Framework
Django>=5.0,<6.0

# Database Drivers
psycopg2-binary>=2.9.9  # PostgreSQL (for production)
# SQLite is built-in with Python (for development)

# Environment Variables
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# AI Integration (OpenAI-compatible API)
openai>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Document Processing
python-docx>=1.0.0  # Word documents
python-pptx>=0.6.21  # PowerPoint documents

# PDF Generation
weasyprint>=60.0
markdown>=3.5.0

# Excel/CSV Processing
openpyxl>=3.1.0
pandas>=2.0.0

# Security
argon2-cffi>=23.1.0

# Production Server
gunicorn>=21.0.0

# Development Tools
django-debug-toolbar>=4.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0

# HTMX Support (optional - for dynamic interactions)
django-htmx>=1.17.0

# Form Rendering
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.1

# Additional Utilities
python-dateutil>=2.8.0
